requirements.txt
django==4.1
djangorestframework==3.14.0
django-cacheops==7.0

psycopg2==2.9.6

celery[redis]==5.2.7
celery-singleton==0.3.1
flower==1.2.0

pyyaml==6.0
uritemplate==4.1.1